[better error msgs] reported dimension spans

+++ src/utils/sourcemap.rs
@@ -340,8 +340,8 @@ impl Span {
         self.lo <= p && self.hi >= p
     }
     /// computes length of the span
-    pub fn len(&self) -> u64 {
-        self.hi.0 - self.lo.0
+    pub fn len(&self) -> usize {
+        (self.hi.0 - self.lo.0) as usize
     }
     /// merges two spans, same as `+` operator
     pub fn merge(self, other: Span) -> Span {
